+++ finagle-base-http/src/main/scala/com/twitter/finagle/http/util/FailingWriter.scala
@@ -1,15 +1,16 @@
 package com.twitter.finagle.http.util
 
-import com.twitter.io.{Buf, StreamTermination, Writer}
+import com.twitter.finagle.http.Chunk
+import com.twitter.io.{StreamTermination, Writer}
 import com.twitter.util.{Future, Time}
 
 /**
  * A [[Writer]] whose each operation fails with [[UnsupportedOperationException]].
  */
-private[finagle] object FailingWriter extends Writer[Buf] {
+private[finagle] object FailingWriter extends Writer[Chunk] {
   def fail(cause: Throwable): Unit = ()
 
-  def write(element: Buf): Future[Unit] =
+  def write(element: Chunk): Future[Unit] =
     Future.exception(new UnsupportedOperationException("FailingWriter"))
 
   def close(deadline: Time): Future[Unit] =