+++ finagle-netty4-http/src/main/scala/com/twitter/finagle/netty4/http/Netty4FormPostEncoder.scala
@@ -110,7 +110,7 @@ private[finagle] object Netty4FormPostEncoder extends FormPostEncoder {
   private[this] def chunkedReqToFinagle(nettyReq: HttpRequest, body: Buf): Request = {
     val req = Bijections.netty.chunkedRequestToFinagle(
       nettyReq,
-      Reader.fromBuf(body),
+      Reader.value(Chunk.apply(body)),
       new InetSocketAddress(0)
     )